+++ src/oscar/Maths/ClosedInterval.h
@@ -1,6 +1,9 @@
 #pragma once
 
+#include <oscar/Utils/Assertions.h>
+
 #include <concepts>
+#include <cstddef>
 
 namespace osc
 {
@@ -8,6 +11,23 @@ namespace osc
     template<typename T>
     requires std::equality_comparable<T> and std::totally_ordered<T>
     struct ClosedInterval final {
+        friend bool operator==(const ClosedInterval&, const ClosedInterval&) = default;
+
+        template<std::integral U>
+        constexpr T step_size(U nsteps) const
+        {
+            if (nsteps <= 1) {
+                return upper - lower;  // edge-case
+            }
+
+            return (upper - lower) / (static_cast<std::make_unsigned_t<U>>(nsteps) - 1);
+        }
+
+        constexpr T normalized_interpolant_at(T v) const
+        {
+            return (v - lower) / (upper - lower);
+        }
+
         T lower;
         T upper;
     };